requirements.txt
anyio==4.9.0
apsw==3.49.1.0
apswutils==0.0.2
beautifulsoup4==4.13.4
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
fastcore==1.8.1
fastlite==0.1.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
mistune==3.1.3
oauthlib==3.2.2
packaging==25.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-fasthtml==0.12.15
python-multipart==0.0.20
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
